Some trivial refactoring.
