+++ AsyncRx.NET/System.Reactive.Async.Subjects/System/Reactive/Subjects/ReplayAsyncSubject.cs
@@ -11,92 +11,6 @@ using System.Threading.Tasks;
 
 namespace System.Reactive.Subjects
 {
-    public sealed class SequentialReplayAsyncSubject<T> : ReplayAsyncSubject<T>
-    {
-        public SequentialReplayAsyncSubject()
-            : base(false)
-        {
-        }
-
-        public SequentialReplayAsyncSubject(int bufferSize)
-            : base(false, bufferSize)
-        {
-        }
-
-        public SequentialReplayAsyncSubject(IAsyncScheduler scheduler)
-            : base(false, scheduler)
-        {
-        }
-
-        public SequentialReplayAsyncSubject(int bufferSize, IAsyncScheduler scheduler)
-            : base(false, bufferSize, scheduler)
-        {
-        }
-
-        public SequentialReplayAsyncSubject(TimeSpan window)
-            : base(false, window)
-        {
-        }
-
-        public SequentialReplayAsyncSubject(TimeSpan window, IAsyncScheduler scheduler)
-            : base(false, window, scheduler)
-        {
-        }
-
-        public SequentialReplayAsyncSubject(int bufferSize, TimeSpan window)
-            : base(false, bufferSize, window)
-        {
-        }
-
-        public SequentialReplayAsyncSubject(int bufferSize, TimeSpan window, IAsyncScheduler scheduler)
-            : base(false, bufferSize, window, scheduler)
-        {
-        }
-    }
-
-    public sealed class ConcurrentReplayAsyncSubject<T> : ReplayAsyncSubject<T>
-    {
-        public ConcurrentReplayAsyncSubject()
-            : base(true)
-        {
-        }
-
-        public ConcurrentReplayAsyncSubject(int bufferSize)
-            : base(true, bufferSize)
-        {
-        }
-
-        public ConcurrentReplayAsyncSubject(IAsyncScheduler scheduler)
-            : base(true, scheduler)
-        {
-        }
-
-        public ConcurrentReplayAsyncSubject(int bufferSize, IAsyncScheduler scheduler)
-            : base(true, bufferSize, scheduler)
-        {
-        }
-
-        public ConcurrentReplayAsyncSubject(TimeSpan window)
-            : base(false, window)
-        {
-        }
-
-        public ConcurrentReplayAsyncSubject(TimeSpan window, IAsyncScheduler scheduler)
-            : base(false, window, scheduler)
-        {
-        }
-
-        public ConcurrentReplayAsyncSubject(int bufferSize, TimeSpan window)
-            : base(false, bufferSize, window)
-        {
-        }
-
-        public ConcurrentReplayAsyncSubject(int bufferSize, TimeSpan window, IAsyncScheduler scheduler)
-            : base(false, bufferSize, window, scheduler)
-        {
-        }
-    }
-
     public abstract class ReplayAsyncSubject<T> : IAsyncSubject<T>
     {
         protected readonly IAsyncSubject<T> _impl;
